[scanner] test: split oversized hook test files
Split web/src/hooks/__tests__/useSearchIndex.categories.test.ts (1095 lines)
into 4 focused files:
- useSearchIndex.categories.basic.test.ts  (§1-14: empty queries, pages, k8s resources, stats, catalog)
- useSearchIndex.categories.limits.test.ts (§15-24: ordering, limits, namespaces, dashboards)
- useSearchIndex.categories.cards.test.ts  (§25-41: placed cards, localStorage, metadata)
- useSearchIndex.categories.edgecases.test.ts (§42-60: edge cases, resource details)

Each file has its own imports, mocks, and beforeEach setup with no shared state.

Fixes #21004
